refactor: move `toDownlodedSurUi` mapper to TilwahMapper
