+++ faith_data/src/commonMain/kotlin/net/thechance/mena/faith/data/mapper/TilwahMapper.kt
@@ -1,7 +1,10 @@
 package net.thechance.mena.faith.data.mapper
 
 import net.thechance.mena.faith.data.database.ReciterDto
+import net.thechance.mena.faith.data.database.SurahAudioDto
 import net.thechance.mena.faith.data.remote.model.tilawah.RecitersRequest
+import net.thechance.mena.faith.domain.entity.Surah
+import net.thechance.mena.faith.domain.model.DownlodedSur
 import net.thechance.mena.faith.domain.model.Reciter
 
 fun RecitersRequest.toReciter(): Reciter = Reciter(
@@ -24,3 +27,13 @@ fun Reciter.toReciterDto(): ReciterDto = ReciterDto(
         nameAr = arabicName,
         tilawahType = tilawahType
     )
+
+fun SurahAudioDto.toDownlodedSurUi(
+    surahName: String,
+    reciterName: List<String>
+): DownlodedSur = DownlodedSur(
+    id = surahId,
+    arabicNameImg = Surah.SurahOrder.entries.first { it.order == surahId },
+    surahName = surahName,
+    recitersName = reciterName
+)
