+++ faith_data/src/commonMain/kotlin/net/thechance/mena/faith/data/repository/QuranRepositoryImpl.kt
@@ -8,6 +8,7 @@ import net.thechance.mena.faith.data.database.SurahAudioDao
 import net.thechance.mena.faith.data.database.SurahAudioDto
 import net.thechance.mena.faith.data.datastore.TilawahDataStore
 import net.thechance.mena.faith.data.mapper.toAyah
+import net.thechance.mena.faith.data.mapper.toDownlodedSurUi
 import net.thechance.mena.faith.data.mapper.toReciter
 import net.thechance.mena.faith.data.mapper.toReciterDto
 import net.thechance.mena.faith.data.mapper.toSurah
@@ -63,16 +64,11 @@ class QuranRepositoryImpl(
         surahId: Int,
         items: List<SurahAudioDto>
     ): DownlodedSur {
-        return DownlodedSur(
-            id = surahId,
-            arabicNameImg = getSurahArabicName(surahId),
+        return items.first().toDownlodedSurUi(
             surahName = getSurahById(surahId).name,
-            recitersName = getRecitersNames(items)
+            reciterName = getRecitersNames(items)
         )
     }
-
-    private fun getSurahArabicName(surahId: Int) =
-        Surah.SurahOrder.entries.first { it.order == surahId }
 
     private suspend fun getRecitersNames(items: List<SurahAudioDto>): List<String> =
         items.map { recitersDao.getReciterById(it.reciterId).name }
